requirements.txt
pandas
numpy
scikit-learn
pyyaml
loguru
great_expectations
mlflow
dvc
dvc[gdrive]  # If you want to use Google Drive as remote storage